Increase client-side timeout when watching (#18)
Motivation
`AbortedStreamException` is raised a lot in Central Dogma server. This happens when a client
sends a request and closes the connection before it gets a response.
I think it happens in the watch API.
When the `Watcher` sends a watch request, it sets the client-side timeout
as original timeout + 1 second. The orginal timeout is used in the server-side. The server responses with
`304` HTTP status when it times out.
If the client gets the `304` response, it will read the response and reuse the connection.
However, the client-side timeout is too short, it closed the connection even before it gets response.
This result in a lot of `AbortedStreamException` in the server.

Modification

- Increase the timeout buffer to 5 seconds

Result

- Fewer exception logs